requirements.txt

streamlit==1.34.0 
openai==1.30.1 
streamlit-extras==0.4.2 
pydantic==1.*
python-dotenv
os
streamlit-float